requirements.txt
Flask==2.3.3
Flask-Limiter==2.9
python-dotenv==1.0.0
